ref: move message filter into tools mod
